fix(blog): stop every post page emitting the whole blog's schema
The Blog and BreadcrumbList JSON-LD lived in the /blog section layout,
which also wraps /blog/[id]. Every one of the 44 post pages therefore
shipped a Blog entity listing all 44 BlogPostings, plus a second
BreadcrumbList that stopped at /blog and contradicted the three-level
one the post itself emits.

Measured on a single post page before this change: 45 BlogPosting
blocks and 2 BreadcrumbLists. That is 44 URLs carrying byte-identical
structured data and two competing breadcrumb trails, which is the
"duplicate, Google chose a different canonical" signal in Search
Console -- no canonical tag outranks structured data telling Google
that 44 pages describe the same 44 articles.

The listing schema moves to the index page, which is the only page it
describes. Post pages now emit exactly one BlogPosting and one
BreadcrumbList. The per-post author in the listing is typed as a Person
now that the byline is a named human.
